first-aid-test/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
